いいねがマイナスにできるバグを修正

+++ app/controllers/likes_controller.rb
@@ -6,18 +6,29 @@ class LikesController < ApplicationController
 
     def create
         @impression = Impression.find_by(id: params[:impression_id])
-        @impression.like_count += 1
-        @impression.update_columns(like_count: @impression.like_count)
-        if signed_in? && params[:user_id].present?
+
+        unless signed_in?
+            @impression.like_count += 1
+            @impression.update_columns(like_count: @impression.like_count)
+        end
+        if signed_in? && params[:user_id].present? && @impression.likes.find_by(user_id: current_user.id).blank?
+            @impression.like_count += 1
+            @impression.update_columns(like_count: @impression.like_count)
             like = Like.create(impression_id: like_params[:impression_id], user_id: like_params[:user_id])
         end
     end
 
     def destroy
         @impression = Impression.find_by(id: params[:impression_id])
-        @impression.like_count -= 1
-        @impression.update_columns(like_count: @impression.like_count)
-        if signed_in? && params[:user_id].present?
+        unless signed_in?
+            if @impression.like_count > 0
+                @impression.like_count -= 1
+                @impression.update_columns(like_count: @impression.like_count)
+            end
+        end
+        if signed_in? && params[:user_id].present? && @impression.likes.find_by(user_id: current_user.id).present? && @impression.like_count > 0
+            @impression.like_count -= 1
+            @impression.update_columns(like_count: @impression.like_count)
             like = Like.find_by(impression_id: like_params[:impression_id], user_id: like_params[:user_id])
             like.destroy!
         end
